// src/pages/Users/types/User.ts
export type User = {
    id: string
    fullName: string
    document: string
    birthDate: Date
    email: string
    emailVerified: boolean
    mobile: string
    zipCode: string
    addressName: string
    number: string
    complement: string
    neighborhood: string
    city: string
    state: string
  }